Move tracker utils into new storage sub-package
The tracker only makes sense in the context of storage. Splitting
up the module also clarifies there were multiple classes within.
