+++ src/snsary/utils/__init__.py
@@ -1,7 +1,7 @@
 from . import scraper
 from .logger import HasLogger, configure_logging, get_logger
 from .service import Service
-from .storage import get_storage
+from .storage import HasStore, MaxTracker, NullTracker, get_storage
 
 __all__ = [
     "logger",
@@ -10,5 +10,8 @@ __all__ = [
     "Service",
     "get_logger",
     "HasLogger",
+    "HasStore",
+    "MaxTracker",
+    "NullTracker",
     "get_storage",
 ]
