+++ src/snsary/utils/storage/__init__.py
@@ -3,6 +3,15 @@ A store has a dict-like interface to get and set data that a consumer wants to p
 
 By default persistent storage is disabled to avoid creating files without warning - the data stored is only held in memory and not written to disk. Set the ``STORAGE_PATH`` environment variable to enable persistent storage to a file e.g. ``export STORAGE_PATH=snsary.db``.
 
+Storage trackers
+================
+
+Trackers can be used to track the values of readings with specified names using persistent storage. New values are individually compared with old using a specified function e.g. ::
+
+    Tracker('id-for-persistent-storage', myreading=max, othername=min)
+
+Call ``update`` when new readings are available. ``on_change`` is called with kwargs ``old`` and ``new`` when one or more tracked values change - set to a different function to subscribe to changes.
+
 Storage internals
 =================
 
@@ -13,57 +22,15 @@ Even persistent storage will expire, to prevent a build up of dead data over tim
 The storage format is not versioned and may break without warning after an upgrade. You will need to manually remove the file at ``STORAGE_PATH`` if this happens.
 """
 
-import atexit
-import os
-import pickle
+from .backend import HasStore, get_storage
+from .max_tracker import MaxTracker
+from .null_tracker import NullTracker
+from .tracker import Tracker
 
-from cachetools import TTLCache, cached
-
-from .logger import get_logger
-
-
-class HasStore:
-    @property
-    def store(self):
-        """
-        Shortcut to calling ``get_storage()`` directly.
-        """
-        return get_storage()
-
-
-@cached({})
-def get_storage():
-    """
-    Returns a memoised dict-like store.
-    """
-    return _get_storage_backend()
-
-
-def _get_storage_backend():
-    path = os.environ.get("STORAGE_PATH", "")
-
-    if not path:
-        get_logger().debug("Using memory store.")
-        return dict()
-
-    get_logger().debug(f"Using store {path}.")
-    ttl = os.environ.get("STORAGE_TTL", 86400)
-
-    get_logger().debug(f"Store TTL set to {ttl}.")
-    return _get_file_backend(path, int(ttl))
-
-
-def _get_file_backend(path, ttl):
-    try:
-        cache = pickle.load(open(path, "rb"))
-    except Exception:
-        get_logger().debug(f"{path} not found.")
-        get_logger().debug("Creating new store.")
-        cache = TTLCache(float("inf"), ttl)
-
-    def atexit_handler():
-        get_logger().debug(f"Writing store {path}.")
-        pickle.dump(cache, open(path, "wb"))
-
-    atexit.register(atexit_handler)
-    return cache
+__all__ = [
+    "HasStore",
+    "get_storage",
+    "MaxTracker",
+    "NullTracker",
+    "Tracker",
+]
